refactor(console): refactor guide pages
