+++ packages/console/src/pages/Applications/components/GuideModal/index.module.scss
